Rewrote the doc funtionality to be mostly VimL
It uses python webbrowser module still, but now in a way that is
compatible with python3 (and python2). since almost everything in here
is based on VimL and the syntax file for processing, it will be much
cleaner to extend with other methnods of browser opening now.

See ticket #11 for more details